+++ billchop-fe/src/backend/models/Bill.ts
